ci: lock in clippy -D warnings across the workspace (--all-targets)
The 'clippy already zero-warning' claim was false: gating the whole
workspace with -D warnings surfaced 20+ pre-existing errors across
tropel-engine, tropel-http, tropel-report, tropel-executor, tropel-wasm,
tropel-input-k6, tropel-x-grpc and tropel-x-websocket. All are fixed:
type-alias the complex map keys (PoolKey, CertClientMap, TagAggregate),
simplify to is_some_and/.or(Some(..))/map(fn), drop unnecessary casts and
redundant guards/closures, collapse nested ifs in the vu_loop iteration
gate (short-circuit semantics preserved), derive Default, and document the
two function-scoped arc_with_non_send_sync allows with the thread-per-core
invariant (WsSession's std mpsc Receiver is !Sync by design). The CI clippy
step now runs --all-targets so lib+bin+test+bench share the same
zero-warning gate; the 3-OS test matrix, fmt check, and release build were
already in place. cargo clippy --workspace --all-targets -- -D warnings is
clean and all touched-crate tests pass, including the k6 e2e, ramping
wall-clock, and 12s Postman end-to-end suites.

+++ crates/tropel-executor/src/scheduler.rs
@@ -583,7 +583,7 @@ impl VUScheduler {
                 // this ramp-up. Stale state from a fully-drained ramp-down is
                 // harmless (remaining == 0 blocks claims; see clear below).
                 let delta = target - current_vus;
-                let step_delay = stage_duration / delta as u32;
+                let step_delay = stage_duration / delta;
                 for _ in 0..delta {
                     let vu_id = current_vus;
                     let handle = run_vu(self.shared_clone(), vu_id);
@@ -598,7 +598,7 @@ impl VUScheduler {
                 //    claims a surplus slot via try_claim_ramp_down at its next
                 //    iteration start.
                 let delta = current_vus - target;
-                let step_delay = stage_duration / delta as u32;
+                let step_delay = stage_duration / delta;
                 for step in 1..=delta {
                     // Interpolate the ramp-down target from current_vus down
                     // to `target`, one unit at a time. Arm EXACTLY ONE surplus
